Change Project Name from 'JOOReports' to 'JODReports'. Change template FreeMarker NameSpace NAME for dynamic image from 'template' to 'jooscript'. Enhance dynamic image filter with image size control (MaxWidth, MaxHeight and Fit).
git-svn-id: https://jodreports.svn.sourceforge.net/svnroot/jodreports/trunk@7 d507b1d9-9123-0410-be71-93f138d682fd

+++ jodreports/src/main/java/net/sf/jooreports/templates/image/ImageSource.java
@@ -6,6 +6,6 @@ import java.io.OutputStream;
 public interface ImageSource {
 
 	public void write(OutputStream outputStream) throws IOException;
-	public double getWidth();
-	public double getHeight();
+	public int getWidth();
+	public int getHeight();
 }
